[#472] Support --verbose flag.
By default, all Cordova commands must be verbose. This is because we
do not know the purpose of the command and many commands simply output
info.

For a set of commands, we have chosen to silence them by default. These
are the noisy commands that are used often - prepare, compile, build,
run, and emulate. Instead of displaying the noisy output, we display
short concise messages.

The user can enable the verbose output of all commands with --verbose.

On any error, the original Cordova output is displayed for debugging.